refactor(core): create separate module for protocol_interfaces
