+++ lib/core.dart
@@ -12,4 +12,5 @@ library core;
 export "src/core/definitions.dart";
 export "src/core/exceptions.dart";
 export "src/core/implementation.dart";
+export "src/core/protocol_interfaces.dart";
 export "src/core/scripting_api.dart";
