+++ lib/src/core/implementation.dart
@@ -12,7 +12,4 @@ export "implementation/codecs/content_codec.dart";
 export "implementation/content.dart";
 export "implementation/content_serdes.dart";
 export "implementation/discovery/discovery_configuration.dart";
-export "implementation/protocol_interfaces/protocol_client.dart";
-export "implementation/protocol_interfaces/protocol_client_factory.dart";
-export "implementation/protocol_interfaces/protocol_server.dart";
 export "implementation/servient.dart" show Servient;
