+++ lib/src/core/protocol_interfaces.dart
@@ -1,0 +1,9 @@
+// Copyright 2024 Contributors to the Eclipse Foundation. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+//
+// SPDX-License-Identifier: BSD-3-Clause
+
+export "protocol_interfaces/protocol_client.dart";
+export "protocol_interfaces/protocol_client_factory.dart";
+export "protocol_interfaces/protocol_server.dart";
